chatbot_poc_backendapi/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.1
sqlalchemy==2.0.20
psycopg2-binary==2.9.9
passlib[argon2]
python-dotenv==1.0.0
alembic==1.12.0
python-jose[cryptography]==3.3.0
openai
fitz
tiktoken
faiss-cpu
slowapi
requests
beautifulsoup4
lxml
tqdm